% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_valid_names.R
\name{as_valid_names}
\alias{as_valid_names}
\title{creation de noms de colonnes valides}
\usage{
as_valid_names(z)
}
\arguments{
\item{z}{le vecteur de noms a traiter}
}
\value{
le vecteur z assaini
}
\description{
la fonction prend en parametre
un vecteur de chaine de caracteres, et la transforme
en supprimant l'ensemble des accents, ponctuation, etc.
}
\examples{
m <- system.file(package = "nettoyage",
   "examples", "m-chanson.txt")
z <- readLines(m, encoding = "UTF-8")
as_valid_names(z)
}
