% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_david.R
\name{summary_david}
\alias{summary_david}
\title{ma version de summary}
\usage{
summary_david(dat)
}
\arguments{
\item{dat}{le dataset a analyser}
}
\description{
une version tellement plus mieux
que celle de la R Core Team.
}
\examples{
summary_david(iris)

if(require("ggplot2")){
  summary_david(ggplot2::diamonds)
}

}
